+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog may include unreleased versions.
 See [version roadmap](https://www.graalvm.org/release-notes/version-roadmap/) for release dates.
 
 ## Version 22.3.0
+* Added an experimental option `--js.unhandled-rejections=handler` that allows to use a custom callback to track unhandled promise rejections.
 * Updated Node.js to version 16.16.0.
 * ECMA-402 Internationalization API has been enabled by default. It can be disabled by `--js.intl-402=false` option.
 * Implemented the [Decorators (stage 3)](https://github.com/tc39/proposal-decorators) proposal.